requirements.txt
# Core dependencies
numpy==1.26.4
meteodata-lab==0.4.0
pyproj==3.7.2
rasterio==1.4.3
scipy==1.16.1
geopy==2.4.1
mcp==1.13.1
python-dotenv==1.1.1
ollama==0.5.3

# Testing dependencies
pytest-asyncio==1.1.0
pytest-mock==3.15.1